refactor(portable): extract OCI protocol and API orchestration (#1021)
