+++ BUILD.bazel
@@ -66,6 +66,7 @@ build_test(
         "//crates/graphforge-filesystem:graphforge_filesystem",
         "//crates/graphforge-cypher:graphforge_cypher",
         "//crates/graphforge-discovery:graphforge_discovery",
+        "//crates/graphforge-portable-oci:graphforge_portable_oci",
         "//crates/graphforge-ir:graphforge_ir",
         "//crates/graphforge-ontology:graphforge_ontology",
         "//crates/graphforge-plan:graphforge_plan",
@@ -84,6 +85,7 @@ test_suite(
         "//crates/graphforge-filesystem:graphforge_filesystem_test",
         "//crates/graphforge-cypher:graphforge_cypher_test",
         "//crates/graphforge-discovery:graphforge_discovery_test",
+        "//crates/graphforge-portable-oci:graphforge_portable_oci_test",
         "//crates/graphforge-ir:graphforge_ir_test",
         "//crates/graphforge-ontology:graphforge_ontology_test",
         "//crates/graphforge-plan:graphforge_plan_test",
@@ -134,6 +136,7 @@ build_test(
         "//crates/graphforge-filesystem:graphforge_filesystem",
         "//crates/graphforge-cypher:graphforge_cypher",
         "//crates/graphforge-discovery:graphforge_discovery",
+        "//crates/graphforge-portable-oci:graphforge_portable_oci",
         "//crates/graphforge-exec:graphforge_exec",
         "//crates/graphforge-io:graphforge_io",
         "//crates/graphforge-ir:graphforge_ir",
