+++ crates/graphforge-portable-oci/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "graphforge-portable-oci"
+description = "GraphForge portable OCI protocol and authenticated transport"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+license-file.workspace = true
+repository.workspace = true
+
+[dependencies]
+graphforge-core = { version = "0.5.2", path = "../graphforge-core" }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sha2 = { workspace = true }
+url = { workspace = true }
+ureq = { workspace = true }
+
+[lints]
+workspace = true
