+++ crates/graphforge-storage/Cargo.toml
@@ -40,7 +40,6 @@ tempfile = "3"
 thiserror = { workspace = true }
 tokio = { workspace = true }
 unicode-normalization = { workspace = true }
-ureq = { workspace = true }
 uuid = { workspace = true }
 
 [target.'cfg(unix)'.dependencies]
