+++ docs/development/bazel-migration-ledger.md
@@ -126,6 +126,7 @@ Authoritative machine-readable map: `tools/bazel/parity/migration_target_map.jso
 | `graphforge-cypher` | `graphforge_cypher` | `lib` | `crates/graphforge-cypher/src/lib.rs` | `//crates/graphforge-cypher:graphforge_cypher` | `mapped` | #10; unit tests `//crates/graphforge-cypher:graphforge_cypher_test` |
 | `graphforge-cypher` | `corpus` | `integration-test` | `crates/graphforge-cypher/tests/corpus.rs` | `//crates/graphforge-cypher:corpus` | `mapped` | #8 |
 | `graphforge-cypher` | `compile` | `bench` | `crates/graphforge-cypher/benches/compile.rs` | — | `exception` | RT-codspeed-bench; CodSpeed divan benchmark |
+| `graphforge-portable-oci` | `graphforge_portable_oci` | `lib` | `crates/graphforge-portable-oci/src/lib.rs` | `//crates/graphforge-portable-oci:graphforge_portable_oci` | `mapped` | #1021; unit tests `//crates/graphforge-portable-oci:graphforge_portable_oci_test` |
 | `graphforge-discovery` | `graphforge_discovery` | `lib` | `crates/graphforge-discovery/src/lib.rs` | `//crates/graphforge-discovery:graphforge_discovery` | `mapped` | #908; unit tests `//crates/graphforge-discovery:graphforge_discovery_test` |
 | `graphforge-discovery` | `contract_artifacts` | `integration-test` | `crates/graphforge-discovery/tests/contract_artifacts.rs` | `//crates/graphforge-discovery:contract_artifacts` | `mapped` | #910; deterministic schema and conformance fixture parity |
 | `graphforge-exec` | `graphforge_exec` | `lib` | `crates/graphforge-exec/src/lib.rs` | `//crates/graphforge-exec:graphforge_exec` | `mapped` | #9; unit tests `//crates/graphforge-exec:graphforge_exec_test` |
