+++ tools/bazel/drift/cargo_feature_fingerprint.json
@@ -1,6 +1,6 @@
 {
   "schema": "graphforge.cargo-feature-fingerprint.v1",
-  "sha256": "1c4d93f6d396946a13a1c50bb25d7bea788d90fd511d3834ab75666644f818ca",
+  "sha256": "1fe9ccbbed1ed043e498f61160e93a4e5f861342e8dc2fbd24e44318ff464bb2",
   "entries": [
     {
       "name": "graphforge-api",
@@ -146,6 +146,15 @@
         },
         {
           "name": "graphforge-ontology",
+          "req": "^0.5.2",
+          "features": [],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "graphforge-portable-oci",
           "req": "^0.5.2",
           "features": [],
           "optional": false,
@@ -1610,6 +1619,73 @@
       ]
     },
     {
+      "name": "graphforge-portable-oci",
+      "version": "0.5.2",
+      "features": [],
+      "dependencies": [
+        {
+          "name": "graphforge-core",
+          "req": "^0.5.2",
+          "features": [],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "serde",
+          "req": "^1",
+          "features": [
+            "derive"
+          ],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "serde_json",
+          "req": "^1",
+          "features": [],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "sha2",
+          "req": "^0.11",
+          "features": [],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "ureq",
+          "req": "=3.4.0",
+          "features": [
+            "rustls"
+          ],
+          "optional": false,
+          "uses_default_features": false,
+          "kind": null,
+          "target": null
+        },
+        {
+          "name": "url",
+          "req": "^2",
+          "features": [
+            "serde"
+          ],
+          "optional": false,
+          "uses_default_features": true,
+          "kind": null,
+          "target": null
+        }
+      ]
+    },
+    {
       "name": "graphforge-provenance",
       "version": "0.5.2",
       "features": [],
@@ -2218,17 +2294,6 @@
           "features": [],
           "optional": false,
           "uses_default_features": true,
-          "kind": null,
-          "target": null
-        },
-        {
-          "name": "ureq",
-          "req": "=3.4.0",
-          "features": [
-            "rustls"
-          ],
-          "optional": false,
-          "uses_default_features": false,
           "kind": null,
           "target": null
         },
